+++ vitest.config.ts
@@ -33,6 +33,9 @@ export default defineConfig(({ command, mode }) => {
         symbolId: '[name]',
       }),
       checker({
+        overlay: {
+          initialIsOpen: false,
+        },
         typescript: true,
         eslint: {
           lintCommand:
@@ -51,8 +54,9 @@ export default defineConfig(({ command, mode }) => {
       preprocessorOptions: {
         scss: {
           additionalData: `
-          @import "@/styles/_mixins.scss";
           @import "@/styles/_functions.scss";
+          @import "@/styles/_placeholders.scss";
+          @import "@/styles/_mixins.scss";
         `,
         },
       },